requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
pdfminer.six==20221105
tqdm==4.65.0
nltk==3.8.1
tiktoken==0.5.2
sentence-transformers
huggingface-hub
chromadb==0.4.22
